fixed interrupt priority configuration not setting

+++ drivers/Inc/stm32wb55rg_gpioDrivers.h
@@ -126,7 +126,7 @@ void GPIO_WriteToOutputPort(GPIO_RegDef_t *pGPIOx, uint8_t Value);
 // IRQ Configuration and IRQ Handling
 
 void GPIO_IRQInterruptConfig(uint8_t IRQNumber, uint8_t ENorDI);
-void GPIO_IRQPriorityConfig(uint8_t IRQNumber, uint8_t IRQPriority);
+void GPIO_IRQPriorityConfig(uint8_t IRQNumber, uint32_t IRQPriority);
 void GPIO_IRQHandling(uint8_t PinNumber);
 
 #endif /* INC_STM32WB55RG_GPIODRIVERS_H_ */
